+++ Source/aweXpect.Reflection/Helpers/ConstructorInfoHelpers.cs
@@ -6,48 +6,43 @@ using aweXpect.Reflection.Collections;
 namespace aweXpect.Reflection.Helpers;
 
 /// <summary>
-///     Extension methods for <see cref="MethodInfo" />.
+///     Extension methods for <see cref="ConstructorInfo" />.
 /// </summary>
-internal static class MethodInfoExtensions
+internal static class ConstructorInfoHelpers
 {
 	/// <summary>
-	///     Checks if the <paramref name="methodInfo" /> has the specified <paramref name="accessModifiers" />.
+	///     Checks if the <paramref name="constructorInfo" /> has the specified <paramref name="accessModifiers" />.
 	/// </summary>
-	/// <param name="methodInfo">The <see cref="MethodInfo" /> which is checked to have the attribute.</param>
+	/// <param name="constructorInfo">The <see cref="ConstructorInfo" /> which is checked to have the attribute.</param>
 	/// <param name="accessModifiers">
 	///     The <see cref="AccessModifiers" />.
 	///     <para />
 	///     Supports specifying multiple <see cref="AccessModifiers" />.
 	/// </param>
 	public static bool HasAccessModifier(
-		this MethodInfo? methodInfo,
+		this ConstructorInfo constructorInfo,
 		AccessModifiers accessModifiers)
 	{
-		if (methodInfo == null)
-		{
-			return false;
-		}
-
 		if (accessModifiers.HasFlag(AccessModifiers.Internal) &&
-		    methodInfo.IsAssembly)
+		    constructorInfo.IsAssembly)
 		{
 			return true;
 		}
 
 		if (accessModifiers.HasFlag(AccessModifiers.Protected) &&
-		    methodInfo.IsFamily)
+		    constructorInfo.IsFamily)
 		{
 			return true;
 		}
 
 		if (accessModifiers.HasFlag(AccessModifiers.Private) &&
-		    methodInfo.IsPrivate)
+		    constructorInfo.IsPrivate)
 		{
 			return true;
 		}
 
 		if (accessModifiers.HasFlag(AccessModifiers.Public) &&
-		    methodInfo.IsPublic)
+		    constructorInfo.IsPublic)
 		{
 			return true;
 		}
@@ -56,27 +51,22 @@ internal static class MethodInfoExtensions
 	}
 
 	/// <summary>
-	///     Checks if the <paramref name="methodInfo" /> has an attribute which satisfies the <paramref name="predicate" />.
+	///     Checks if the <paramref name="constructorInfo" /> has an attribute which satisfies the
+	///     <paramref name="predicate" />.
 	/// </summary>
 	/// <typeparam name="TAttribute">The type of the <see cref="Attribute" />.</typeparam>
-	/// <param name="methodInfo">The <see cref="MethodInfo" /> which is checked to have the attribute.</param>
+	/// <param name="constructorInfo">The <see cref="ConstructorInfo" /> which is checked to have the attribute.</param>
 	/// <param name="predicate">
 	///     (optional) A predicate to check the attribute values.
 	///     <para />
 	///     If not set (<see langword="null" />), will only check if the attribute is present.
 	/// </param>
-	/// <param name="inherit">
-	///     <see langword="true" /> to search the inheritance chain to find the attributes; otherwise,
-	///     <see langword="false" />.<br />
-	///     Defaults to <see langword="true" />
-	/// </param>
 	public static bool HasAttribute<TAttribute>(
-		this MethodInfo methodInfo,
-		Func<TAttribute, bool>? predicate = null,
-		bool inherit = true)
+		this ConstructorInfo constructorInfo,
+		Func<TAttribute, bool>? predicate = null)
 		where TAttribute : Attribute
 	{
-		object? attribute = Attribute.GetCustomAttributes(methodInfo, typeof(TAttribute), inherit)
+		object? attribute = constructorInfo.GetCustomAttributes(typeof(TAttribute))
 			.FirstOrDefault();
 		if (attribute is TAttribute castedAttribute)
 		{
