+++ Source/aweXpect.Reflection/Helpers/MethodInfoHelpers.cs
@@ -6,29 +6,60 @@ using aweXpect.Reflection.Collections;
 namespace aweXpect.Reflection.Helpers;
 
 /// <summary>
-///     Extension events for <see cref="EventInfo" />.
+///     Extension methods for <see cref="MethodInfo" />.
 /// </summary>
-internal static class EventInfoExtensions
+internal static class MethodInfoHelpers
 {
 	/// <summary>
-	///     Checks if the <paramref name="eventInfo" /> has the specified <paramref name="accessModifiers" />.
+	///     Checks if the <paramref name="methodInfo" /> has the specified <paramref name="accessModifiers" />.
 	/// </summary>
-	/// <param name="eventInfo">The <see cref="FieldInfo" /> which is checked to have the attribute.</param>
+	/// <param name="methodInfo">The <see cref="MethodInfo" /> which is checked to have the attribute.</param>
 	/// <param name="accessModifiers">
 	///     The <see cref="AccessModifiers" />.
 	///     <para />
 	///     Supports specifying multiple <see cref="AccessModifiers" />.
 	/// </param>
 	public static bool HasAccessModifier(
-		this EventInfo? eventInfo,
+		this MethodInfo? methodInfo,
 		AccessModifiers accessModifiers)
-		=> eventInfo?.AddMethod.HasAccessModifier(accessModifiers) == true;
+	{
+		if (methodInfo == null)
+		{
+			return false;
+		}
+
+		if (accessModifiers.HasFlag(AccessModifiers.Internal) &&
+		    methodInfo.IsAssembly)
+		{
+			return true;
+		}
+
+		if (accessModifiers.HasFlag(AccessModifiers.Protected) &&
+		    methodInfo.IsFamily)
+		{
+			return true;
+		}
+
+		if (accessModifiers.HasFlag(AccessModifiers.Private) &&
+		    methodInfo.IsPrivate)
+		{
+			return true;
+		}
+
+		if (accessModifiers.HasFlag(AccessModifiers.Public) &&
+		    methodInfo.IsPublic)
+		{
+			return true;
+		}
+
+		return false;
+	}
 
 	/// <summary>
-	///     Checks if the <paramref name="eventInfo" /> has an attribute which satisfies the <paramref name="predicate" />.
+	///     Checks if the <paramref name="methodInfo" /> has an attribute which satisfies the <paramref name="predicate" />.
 	/// </summary>
 	/// <typeparam name="TAttribute">The type of the <see cref="Attribute" />.</typeparam>
-	/// <param name="eventInfo">The <see cref="EventInfo" /> which is checked to have the attribute.</param>
+	/// <param name="methodInfo">The <see cref="MethodInfo" /> which is checked to have the attribute.</param>
 	/// <param name="predicate">
 	///     (optional) A predicate to check the attribute values.
 	///     <para />
@@ -40,12 +71,12 @@ internal static class EventInfoExtensions
 	///     Defaults to <see langword="true" />
 	/// </param>
 	public static bool HasAttribute<TAttribute>(
-		this EventInfo eventInfo,
+		this MethodInfo methodInfo,
 		Func<TAttribute, bool>? predicate = null,
 		bool inherit = true)
 		where TAttribute : Attribute
 	{
-		object? attribute = Attribute.GetCustomAttributes(eventInfo, typeof(TAttribute), inherit)
+		object? attribute = Attribute.GetCustomAttributes(methodInfo, typeof(TAttribute), inherit)
 			.FirstOrDefault();
 		if (attribute is TAttribute castedAttribute)
 		{
