+++ Source/aweXpect.Reflection/Helpers/ThatHelpers.cs
@@ -4,7 +4,7 @@ using aweXpect.Core;
 
 namespace aweXpect.Reflection.Helpers;
 
-internal static class ThatExtensions
+internal static class ThatHelpers
 {
 	[ExcludeFromCodeCoverage]
 	public static IExpectThat<T> Get<T>(this IThat<T> subject)
